finishing core library

+++ FastShare.Core/Model/FastShareFileInfo.cs
@@ -7,6 +7,6 @@ namespace FastShare.Core.Model
     public class FastShareFileInfo
     {
         public string Title { get; internal set; }
-        public int Length { get; internal set; }
+        public long Length { get; internal set; }
     }
 }
